src/lambda-payments-executor/requirements.txt
aws-lambda-powertools>=3.22.0
boto3>=1.40.55
pydantic>=2.12.3
requests>=2.32.5
opentelemetry-api>=1.38.0
opentelemetry-sdk>=1.38.0
opentelemetry-exporter-otlp-proto-http>=1.38.0